created seperate module for items page. seperated previous code from the home module
