+++ wishlist/home/urls.py
@@ -3,5 +3,4 @@ from . import views
 
 urlpatterns = [
     path('', views.home, name='home-page'),
-    path('items/', views.ListAllItems, name='items-page')
 ]
