+++ wishlist/home/views.py
@@ -15,30 +15,9 @@ posts = [
     }
 ]
 
-items = [
-    {
-        'item_no' : '1',
-        'item_name' : 'Water Bottle',
-        'content' : 'drink water duh',
-    },
-    {
-        'item_no' : '2',
-        'item_name' : 'Phone',
-        'content' : 'its a phone, what did you expect?',
-    }
-]
-
 # Create your views here.
 def home(request):
     context = {
         'posts' : posts
     }
     return render(request, 'home/home.html', context)
-
-
-def ListAllItems(request):
-    # pull from db
-    context = {
-        'items' : items
-    }
-    return render(request, 'home/items.html', context)
