+++ wishlist/wishlist/urls.py
@@ -15,9 +15,10 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path, include
+from items import views as item_views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('home.urls')),
-    path('items/', include('home.urls'))
+    path('items/', item_views.ListAllItems)
 ]
